Move logic into knot-ide crate
