+++ crates/red_knot_ide/src/find_node.rs
@@ -5,8 +5,6 @@ use ruff_python_ast::{AnyNodeRef, Keyword};
 use ruff_text_size::{Ranged, TextRange};
 use std::fmt;
 use std::fmt::Formatter;
-// What we want is a `find_node_at` function that returns the node that fully encloses offset.
-// It may return two nodes if the offset falls exactly between two nodes.
 
 /// Returns the node with a minimal range that fully contains `range`.
 ///
@@ -124,10 +122,6 @@ impl<'a> CoveringNode<'a> {
 
     pub(crate) fn parent(&self) -> Option<AnyNodeRef<'a>> {
         self.ancestors.last().copied()
-    }
-
-    pub(crate) fn ancestors(&self) -> impl DoubleEndedIterator<Item = AnyNodeRef<'a>> + '_ {
-        std::iter::once(self.node).chain(self.ancestors.iter().rev().copied())
     }
 
     /// Finds the first node that fully covers the range and fulfills the given predicate.
