+++ crates/red_knot_python_semantic/src/semantic_index/symbol.rs
@@ -118,7 +118,7 @@ impl<'db> ScopeId<'db> {
         self.node(db).scope_kind().is_type_parameter()
     }
 
-    pub fn node(self, db: &dyn Db) -> &NodeWithScopeKind {
+    pub(crate) fn node(self, db: &dyn Db) -> &NodeWithScopeKind {
         self.scope(db).node()
     }
