+++ crates/red_knot_server/Cargo.toml
@@ -11,13 +11,13 @@ repository = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
+red_knot_ide = { workspace = true }
 red_knot_project = { workspace = true }
 red_knot_python_semantic = { workspace = true }
 
 ruff_db = { workspace = true, features = ["os"] }
 ruff_notebook = { workspace = true }
 ruff_python_ast = { workspace = true }
-ruff_python_parser = { workspace = true }
 ruff_source_file = { workspace = true }
 ruff_text_size = { workspace = true }
 
@@ -31,14 +31,11 @@ salsa = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 shellexpand = { workspace = true }
-smallvec = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 
 [dev-dependencies]
 red_knot_vendored = { workspace = true }
-
-insta = { workspace = true }
 
 [target.'cfg(target_vendor = "apple")'.dependencies]
 libc = { workspace = true }
