--- crates/red_knot_server/src/semantic.rs
@@ -1,1 +1,0 @@
-pub(crate) mod goto;
